requirements.txt
#python==3.10
streamlit==1.41.1
requests==2.32.3
PyPDF2==3.0.1
pandas==2.2.3
python-dotenv==1.0.1
groq==0.13.1
